+++ src/WheeledRobot.cpp
@@ -4,29 +4,29 @@ WheeledRobot::WheeledRobot() {
     m_status.status = actionlib_msgs::GoalStatus::ACTIVE;
 }
 
-void WheeledRobot::setGoal(std::vector<Vector3D<float>> t_goal) {
+void WheeledRobot::setGoal(std::vector<Vector3D<double>> t_goal) {
     m_GoalPose.clear();
     m_GoalPose = t_goal;
 }
 
-void WheeledRobot::setGoal(Vector3D<float> t_goal) {
+void WheeledRobot::setGoal(Vector3D<double> t_goal) {
     m_GoalPose.clear();
     m_GoalPose.push_back(t_goal);
 }
 
-Vector2D<float> WheeledRobot::getGoalPosition() {
+Vector2D<double> WheeledRobot::getGoalPosition() {
     return m_GoalPose.back().project_xy();
 }
 
-float WheeledRobot::getGoalHeading() {
+double WheeledRobot::getGoalHeading() {
     return m_GoalPose.back().z;
 }
 
-Vector2D<float> WheeledRobot::getCurrentPosition() {
+Vector2D<double> WheeledRobot::getCurrentPosition() {
     return m_CurrentPosition;
 }
 
-float WheeledRobot::getCurrentHeading() {
+double WheeledRobot::getCurrentHeading() {
     return m_CurrentHeading;
 }
 
@@ -74,7 +74,7 @@ bool WheeledRobot::clearQueue() {
         Vector2DMsg t_GoalPosMsg;
         t_GoalPosMsg.data = m_queue.front().project_xy();
         Vector3DMessage t_GoalHeadingMsg;
-        t_GoalHeadingMsg.setVector3DMessage(Vector3D<float>({0,0, m_queue.front().z}));
+        t_GoalHeadingMsg.setVector3DMessage(Vector3D<double>({0,0, m_queue.front().z}));
         this->emit_message((DataMessage*) &t_GoalPosMsg);
         this->emit_message((DataMessage*) &t_GoalHeadingMsg);
     }
